requirements.txt
adbutils==1.0.8
airtest==1.2.0
apkutils2==1.0.0
asgiref==3.4.1
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.9.3
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.9
cigam==0.0.3
click==8.1.3
colored==1.4.3
cryptography==37.0.4
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
Django==3.2.10
django-allauth==0.53.1
django-ckeditor==6.3.2
django-js-asset==2.0.0
django-model-utils==4.3.1
django-mptt==0.14.0
django-notifications==0.1.dev0
django-notifications-hq==1.7.0
django-taggit==3.1.0
facebook-wda==1.4.6
filelock==3.8.0
fire==0.4.0
Flask==2.1.2
Flask-SocketIO==4.3.1
h11==0.13.0
hrpc==1.0.9
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.0.1
jsonfield==3.1.0
loguru==0.6.0
logzero==1.7.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.0.1
mss==4.0.3
numpy==1.19.3
oauthlib==3.2.2
opencv-contrib-python==4.5.5.62
outcome==1.2.0
packaging==20.9
Pillow==8.4.0
pocoui==1.0.82
progress==1.6
py==1.11.0
pycparser==2.21
pyelftools==0.29
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-engineio==3.13.2
python-socketio==4.6.0
python-xlib==0.31
python3-openid==3.2.0
pytz==2021.3
pywinauto==0.6.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
retry==0.9.2
selenium==4.4.0
simple-tornado==0.2.2
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
solox==2.0.6
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
swapper==1.3.0
tabulate==0.8.10
termcolor==1.1.0
tidevice==0.9.7
tornado==6.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.0.1
uiautomator2==2.16.19
urllib3==1.26.7
websocket-client==0.48.0
Werkzeug==2.0.3
whichcraft==0.6.1
wrapt==1.13.3
wsproto==1.1.0
xmltodict==0.13.0
zipp==3.8.0
